Removing partitions and adding foreign keys to tables
