+++ backend/cloud_inquisitor/schema/__init__.py
@@ -1,7 +1,7 @@
-from .base import (BaseModelMixin, Account, Tag, LogEvent, Email, ConfigNamespace, ConfigItem, Role, User,
+from .base import (BaseModelMixin, Account, LogEvent, Email, ConfigNamespace, ConfigItem, Role, User,
                    UserRole, AuditLog, SchedulerBatch, SchedulerJob)
 from .issues import IssueType, IssueProperty, Issue
-from .resource import ResourceType, ResourceProperty, Resource, ResourceMapping
+from .resource import Tag, ResourceType, ResourceProperty, Resource, ResourceMapping
 
 __all__ = (
     'ResourceType', 'ResourceProperty', 'Resource', 'ResourceMapping', 'BaseModelMixin', 'Account', 'Tag', 'LogEvent',
